src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Root Folder of Realtime Doc Context MCP"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.49.0",
    "beautifulsoup4>=4.13.4",
    "duckduckgo-search>=8.0.0",
    "gradio>=5.25.2",
    "mcp[cli]>=1.6.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "sentence-transformers>=4.1.0",
]
